string_view

+++ tools/filesystem.c
@@ -53,8 +53,8 @@ void DeleteAll(const string &dirname) {
   fs::path pDir(dirname);
   std::error_code ec;
   std::uintmax_t n = fs::remove_all(pDir, ec);
-  if (ec.value() != 0) esyslog("tvscraper: ERROR \"%s\", code %i  deleted  \"%s\", %u files", ec.message().c_str(), ec.value(), dirname.c_str(), (unsigned int)n);
-   else if (config.enableDebug) esyslog("tvscraper: deleted  \"%s\", %lu files", dirname.c_str(), n);
+  if (ec.value() != 0) esyslog("tvscraper: ERROR \"%s\", code %i  deleted  \"%s\", %ju files", ec.message().c_str(), ec.value(), dirname.c_str(), n);
+  else if (config.enableDebug) esyslog("tvscraper: deleted  \"%s\", %ju files", dirname.c_str(), n);
 }
 
 bool Download(const std::string &url, const std::string &localPath) {
